# swagger.yaml
host: "localhost:8080"
schemes:
    - http
securityDefinitions:
    Bearer:
        type: apiKey
        name: Authorization
        in: header
        description: enter the token with the Bearer prefix e.g. Bearer eyJhbGci...
definitions:
...
    movie:
        properties:
            description:
                description: The description of the movie.
                example: A computer hacker learns from mysterious rebels about the true nature of his reality and his role in the war against its controllers.
                type: string
            id:
                description: The ID of the movie.
                example: 123
                format: int32
                type: integer
            name:
                description: The name of the movie.
                example: The Matrix
                type: string
            rating:
                description: The rating of the movie.
                example: "8.7"
                type: string
            release_date:
                description: The release date of the movie.
                example: "2012-07-16T00:00:00Z"
                format: date-time
                type: string
        type: object
        title: movieResponse represents the response for a movie.
    allMovies:
        type: array
        items:
            $ref: '#/definitions/movie'
        title: allMoviesResponse represents the response body for all movies.
...
    createMovieRequest:
        type: object
        required: 
            - name
            - description
            - release_date
            - rating
        properties:
            name:
                description: The name of the movie.
                example: Inception
                type: string
            description:
                description: The description of the movie.
                example: A mind-bending action thriller
                type: string
            release_date:
                description: The release date of the movie.
                format: date
                example: "2012-07-16T00:00:00Z"
            rating:
                description: The rating of the movie.
                example: "8.8"
                type: string
        title: userRequest represents the request body for a user.
info: {}
paths:
...
    /movie/create:
        post:
            security:
                - Bearer: []
            operationId: createMovie
            consumes:
                - application/json
            parameters:
                - in: body
                  name: movie
                  schema:
                    $ref: '#/definitions/createMovieRequest'
            produces:
                - application/json
            summary: Creates a new movie with the provided details, ONLY FOR ADMINISTRATORS.
            tags:
                - movies
            responses:
                200:
                    $ref: '#/responses/movieResponse'
                400:
                    $ref: '#/responses/error400Response'
                403:
                    $ref: '#/responses/error403Response'
                404:
                    $ref: '#/responses/error404Response'
                500:
                    $ref: '#/responses/error500Response'
    /movie/delete:
        delete:
            security:
                - Bearer: []
            operationId: deleteMovie
            parameters:
                  - in: query
                    name: id
                    required: true
            produces:
                - application/json
            summary: Delete a movie with the provided ID, ONLY FOR ADMINISTRATORS.
            responses:
                200:
                    schema:
                        type: object
                        properties:
                            id:
                                type: integer
                                format: int32
                                example: 1
                400:
                    $ref: '#/responses/error400Response'
                403:
                    $ref: '#/responses/error403Response'
                404:
                    $ref: '#/responses/error404Response'
            tags:
                - movies
    /movie/update:
        patch:
            security:
                - Bearer: []
            operationId: updateMovie
            consumes:
                - application/json
            summary: Updates a movie with the provided details, ONLY FOR ADMINISTRATORS.
            parameters:
                - in: body
                  name: movie
                  schema:
                    $ref: '#/definitions/movie'
            produces:
                - application/json
            responses:
                200:
                    $ref: '#/responses/movieResponse'
                400:
                    $ref: '#/responses/error400Response'
                403:
                    $ref: '#/responses/error403Response'
                404:
                    $ref: '#/responses/error404Response'
                500:
                    $ref: '#/responses/error500Response'

            tags:
                - movies
    /movies:
        get:
            security:
                - Bearer: []
            operationId: moviesSortedByRating
            summary: Retrieves movies sorted by rating.
            tags:
                - movies
            responses:
                200:
                    $ref: '#/responses/allMoviesRespose'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
    /movies/by-actor-fragment:
        get:
            security:
                - Bearer: []
            consumes:
                - application/json
            operationId: moviesByActorFragment
            parameters:
                - in: query
                  name: actor
                  required: true
                  description: The actor fragment to search for.
            produces:
                - application/json
            summary: Retrieves movies by actor fragment.
            tags:
                - movies
            responses:
                200:
                    $ref: '#/responses/allMoviesRespose'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
    /movies/by-date:
        get:
            security:
                - Bearer: []
            operationId: moviesSortedByReleaseDate
            summary: Retrieves movies sorted by release date.
            tags:
                - movies
            responses:
                200:
                    $ref: '#/responses/allMoviesRespose'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
    /movies/by-name:
        get:
            security:
                - Bearer: []
            operationId: moviesSortedByName
            summary: Retrieves movies sorted by name.
            tags:
                - movies
            responses:
                200:
                    $ref: '#/responses/allMoviesRespose'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
    /movies/by-name-fragment:
        get:
            security:
                - Bearer: []
            consumes:
                - application/json
            operationId: moviesByNameFragment
            parameters:
                - in: query
                  name: name
                  required: true
                  description: The movie fragment to search for.
            produces:
                - application/json
            summary: Retrieves movies by name fragment.
            tags:
                - movies
            responses:
                200:
                    $ref: '#/responses/allMoviesRespose'
                401:
                    $ref: '#/responses/error401Response'
                500:
                    $ref: '#/responses/error500Response'
responses:
...
    allMoviesRespose:
        description: allMoviesResponse represents the response body for all movies.
        properties:
            movies:
                description: The list of movies.
        schema:
            $ref: '#/definitions/allMovies'
    movieResponse:
        properties:
            movie:
                description: movieResponse represents the response for a movie.
        schema:
            $ref: '#/definitions/movie'
...
    error400Response:
        description: Bad Request.
    error401Response:
        description: Unauthorized.
    error403Response:
        description: Forbidden.
    error404Response:
        description: Not Found.
    error500Response:
        description: Internal Server Error.
swagger: "2.0"